move debug analytics configuration into Dagger
