+++ app/src/debug/java/org/cru/godtools/DebugGodToolsApplication.kt
@@ -2,11 +2,9 @@ package org.cru.godtools
 
 import android.content.Context
 import androidx.multidex.MultiDex
-import com.adobe.mobile.Config
 import leakcanary.LeakCanary
 import org.ccci.gto.android.common.leakcanary.crashlytics.CrashlyticsOnHeapAnalyzedListener
 import org.ccci.gto.android.common.leakcanary.timber.TimberSharkLog
-import org.cru.godtools.analytics.TimberAnalyticsService
 import org.cru.godtools.dagger.ApplicationModule
 import org.cru.godtools.dagger.DaggerDebugApplicationComponent
 import shark.SharkLog
@@ -32,14 +30,6 @@ class DebugGodToolsApplication : GodToolsApplication() {
         LeakCanary.config = LeakCanary.config.copy(
             onHeapAnalyzedListener = CrashlyticsOnHeapAnalyzedListener
         )
-    }
-
-    override fun configureAnalyticsServices() {
-        super.configureAnalyticsServices()
-
-        // enable debug logging for various Analytics Services
-        Config.setDebugLogging(true)
-        TimberAnalyticsService.getInstance(null)
     }
 
     private fun initTimber() {
