+++ app/src/debug/java/org/cru/godtools/dagger/DebugApplicationModule.kt
@@ -2,5 +2,5 @@ package org.cru.godtools.dagger
 
 import dagger.Module
 
-@Module(includes = [FlipperModule::class])
+@Module(includes = [DebugServicesModule::class])
 abstract class DebugApplicationModule
