+++ app/src/debug/java/org/cru/godtools/dagger/DebugServicesModule.kt
@@ -1,0 +1,7 @@
+package org.cru.godtools.dagger
+
+import dagger.Module
+import org.cru.godtools.analytics.DebugAnalyticsModule
+
+@Module(includes = [DebugAnalyticsModule::class, FlipperModule::class])
+abstract class DebugServicesModule
